fix(web,protocol): coerce NUMERIC-as-string columns in the replay archive read
Neon's serverless driver returns NUMERIC/DECIMAL columns (starting_price,
winning_price) as JS strings by default. archivedAuctionSchema required
plain z.number(), which does not coerce, so every genuine archived row
failed validation and fetchArchivedAuction's catch turned that into
null -- every completed auction would 404 on replay in production,
despite all-green tests, because the mocks only ever supplied number
literals. Add ::float casts to the read query (matching leaderboard.ts's
existing precedent on the same table) and switch the schema to
z.coerce.number().finite() as a second, cast-independent layer; add
tests that feed string values through the mock. Also broadens the
determinism test to compare the whole winner object, not just amount.

+++ apps/web/src/lib/replay.ts
@@ -62,12 +62,17 @@ export async function fetchArchivedAuction(roomId: string): Promise<ArchivedAuct
 
   try {
     const sql = neon(url);
+    // starting_price/winning_price are NUMERIC (schema.sql); cast to
+    // float here, matching leaderboard.ts's existing precedent on this
+    // same table, so the common case hands archivedAuctionSchema an
+    // actual number rather than relying solely on its z.coerce.number()
+    // fallback (see that schema's comment for why both layers exist).
     const rows = await sql`
-      SELECT item_name       AS "itemName",
-             starting_price  AS "startingPrice",
-             winning_price   AS "winningPrice",
-             winner_nickname AS "winnerNickname",
-             bid_log         AS "bidLog"
+      SELECT item_name              AS "itemName",
+             starting_price::float  AS "startingPrice",
+             winning_price::float   AS "winningPrice",
+             winner_nickname        AS "winnerNickname",
+             bid_log                AS "bidLog"
       FROM completed_auctions
       WHERE room_id = ${roomId}
       LIMIT 1
